refactor(shared): host identical entry-handler methods in SharedEntryHandlerMixin (#486)
AST comparison of the backtest and live execution handlers found exactly
three byte-identical method pairs: _extract_entry_plan, _apply_dynamic_risk,
and get_dynamic_risk_adjustments. Move them verbatim into a shared mixin so
backtest-live parity for entry-plan extraction and dynamic-risk sizing holds
by construction instead of by review.

The remaining same-named methods (process_runtime_decision, execute_entry,
exit-condition checks) genuinely diverge — different return types, snapshot
models, SL/TP units, and live-only validation — so they are intentionally
left engine-specific rather than force-merged.
